requirements.txt
numpy==1.16.3
pandas==0.23.0
pykakasi==1.2
PyYAML==3.12
scipy==1.1.0